% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritss.R
\name{ritss}
\alias{ritss}
\title{RITSS}
\usage{
ritss(
  y,
  x,
  e,
  z,
  indices,
  cut_off_p_value = 0.05,
  screening_function = screening_subprs,
  indices_env_factors = 1,
  verbose = FALSE
)
}
\arguments{
\item{y}{phenotype values.}

\item{x}{genotype data (can be expected counts).}

\item{e}{environmental factors.}

\item{z}{additional covariates. For example, genetic principal components.}

\item{indices}{partitioning of the n samples into K non-overlapping sub samples}

\item{cut_off_p_value}{cut off p-value for the inclusion of scores. The first score is always tested.}

\item{screening_function}{function that performs the screening step.}

\item{indices_env_factors}{indices of environmental factors that are considered in the screening step.}

\item{verbose}{default is FALSE. If TRUE, RITSS outputs more details.}
}
\value{
list of objects including RITSS p-value and corresponding z-score.
}
\description{
RITSS
}
\examples{
n=15000
m=50; d=5; p=5;
y=rnorm(n); x=matrix(rnorm(n*m), nrow=n, ncol=m); 
e=matrix(rnorm(n*d), nrow=n, ncol=d); z=matrix(rnorm(n*p), nrow=n, ncol=p);
indices=create_splits(K=3, n=n, split_ratio=(1/3, 1/3, 1/3))
ritss(y=y, x=x, e=e, z=z, indices=indices)
}
