% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritss.R
\name{screening_subprs}
\alias{screening_subprs}
\title{screening_subprs}
\usage{
screening_subprs(
  y,
  x,
  e,
  z,
  indices_env_factors,
  num_iterations = 10,
  verbose = FALSE
)
}
\description{
screening_subprs
}
