% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritss.R
\name{screening_sv}
\alias{screening_sv}
\title{screening_sv}
\usage{
screening_sv(
  y,
  x,
  e,
  z,
  indices_env_factors,
  fdr_cutoff1 = 0.05,
  fdr_cutoff2 = 0.1,
  verbose = FALSE
)
}
\description{
screening_sv
}
