% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritss.R
\name{ritss}
\alias{ritss}
\title{RITSS}
\usage{
ritss(
  y,
  x,
  e,
  z,
  ii1,
  ii2,
  ii3,
  cut_off_p_value = 0.05,
  screening_function = screening_subprs,
  indices_env_factors = 1,
  verbose = FALSE
)
}
\arguments{
\item{y}{phenotype values.}

\item{x}{genotype data (can be expected counts).}

\item{e}{environmental factors.}

\item{z}{additional covariates. For example, genetic principal components.}

\item{ii1}{indices for I_1.}

\item{ii2}{indices for I_2.}

\item{ii3}{indices for I_3. ii1, ii2, and ii3 should NOT overlap.}

\item{cut_off_p_value}{cut off p-value for the inclusion of scores. The first score is always tested.}

\item{screening_function}{function that performs the screening step.}

\item{indices_env_factors}{indices of environmental factors that are considered in the screening step.}

\item{verbose}{default is FALSE. If TRUE, RITSS outputs more details.}
}
\value{
list of objects including RITSS p-value, corresponding z-score. Also individual z-scores of the three substatistics, the three interaction scores,
and information about the selected variants. z_1, z_2, and z_3 are the z-scores of the three sub statistics. Ui_1 and snps_in_score_i1 correspond to screening step that was performed in I_1 and tested in I_3.
}
\description{
RITSS
}
\examples{
n=15000
m=50; d=5; p=5;
y=rnorm(n); x=matrix(rnorm(n*m), nrow=n, ncol=m); 
e=matrix(rnorm(n*d), nrow=n, ncol=d); z=matrix(rnorm(n*p), nrow=n, ncol=p);
ii1=1:5000; ii2=5001:10000; ii3=10001:15000;
ritss(y, x, e, z, ii1, ii2, ii3)
}
